exported functions
