+++ src/extracts/src/process_watcher/process_manager.rs
@@ -1,9 +1,10 @@
+use crate::process_watcher::extract_process_data::ExtractProcessData;
 use crate::process_watcher::{ProcessState, ProcessWatcher};
 use chrono::{DateTime, Utc};
+use itertools::Itertools;
 use std::collections::{HashMap, HashSet};
 use std::path::Path;
 use std::sync::Arc;
-use itertools::Itertools;
 use sysinfo::{Pid, Process, ProcessRefreshKind, ProcessStatus, System};
 use tokio::sync::{RwLock, RwLockReadGuard, RwLockWriteGuard};
 use tracer_common::recorder::LogRecorder;
@@ -12,7 +13,9 @@ use tracer_common::target_process::{Target, TargetMatchable};
 use tracer_common::types::ebpf_trigger::{
     ExitReason, OutOfMemoryTrigger, ProcessEndTrigger, ProcessStartTrigger,
 };
-use tracer_common::types::event::attributes::process::{CompletedProcess, FullProcessProperties, ProcessProperties, ShortProcessProperties};
+use tracer_common::types::event::attributes::process::{
+    CompletedProcess, FullProcessProperties, ProcessProperties, ShortProcessProperties,
+};
 use tracer_common::types::event::attributes::EventAttributes;
 use tracer_common::types::event::ProcessStatus as TracerProcessStatus;
 use tracing::{debug, error};
@@ -28,24 +31,6 @@ pub struct ProcessManager {
     system: Arc<RwLock<System>>,
 }
 
-fn process_status_to_string(status: &ProcessStatus) -> String {
-    match status {
-        ProcessStatus::Run => "Run".to_string(),
-        ProcessStatus::Sleep => "Sleep".to_string(),
-        ProcessStatus::Idle => "Idle".to_string(),
-        ProcessStatus::Zombie => "Zombie".to_string(),
-        ProcessStatus::Stop => "Stop".to_string(),
-        ProcessStatus::Parked => "Parked".to_string(),
-        ProcessStatus::Tracing => "Tracing".to_string(),
-        ProcessStatus::Dead => "Dead".to_string(),
-        ProcessStatus::UninterruptibleDiskSleep => "Uninterruptible Disk Sleep".to_string(),
-        ProcessStatus::Waking => "Waking".to_string(),
-        ProcessStatus::LockBlocked => "Lock Blocked".to_string(),
-        _ => "Unknown".to_string(),
-    }
-}
-
-
 impl ProcessManager {
     pub fn new(target_manager: TargetManager, log_recorder: LogRecorder) -> Self {
         let state = Arc::new(RwLock::new(ProcessState {
@@ -55,13 +40,13 @@ impl ProcessManager {
             ebpf_task: None,
             oom_victims: HashMap::new(),
         }));
-        
+
         let system = Arc::new(RwLock::new(System::new_all()));
 
         ProcessManager {
             state,
             log_recorder,
-            system
+            system,
         }
     }
 
@@ -342,7 +327,7 @@ impl ProcessManager {
                 ProcessRefreshKind::everything(), // TODO(ENG-336): minimize data collected for performance
             );
         })
-            .await?;
+        .await?;
 
         Ok(())
     }
@@ -510,12 +495,12 @@ impl ProcessManager {
 
             match system.process(process.pid.into()) {
                 Some(system_process) => {
-                    self.gather_process_data(
+                    ExtractProcessData::gather_process_data(
                         system_process,
                         display_name.clone(),
                         process.started_at,
                     )
-                        .await
+                    .await
                 }
                 None => {
                     debug!("Process({}) wasn't found", process.pid);
@@ -578,12 +563,12 @@ impl ProcessManager {
             );
 
             // Don't process input files for update events
-            self.gather_process_data(
+            ExtractProcessData::gather_process_data(
                 system_process,
                 display_name.clone(),
                 process.started_at,
             )
-                .await
+            .await
         };
 
         debug!("Process data completed. PID={}", process.pid);
@@ -598,72 +583,6 @@ impl ProcessManager {
             .await?;
 
         Ok(ProcessResult::Found)
-    }
-
-    pub async fn gather_process_data(
-        &self,
-        proc: &Process,
-        display_name: String,
-        process_start_time: DateTime<Utc>,
-    ) -> ProcessProperties {
-        debug!("Gathering process data for {}", display_name);
-
-        let (container_id, job_id, trace_id) = Self::extract_process_environment_variables(proc);
-
-        let working_directory = proc.cwd().map(|p| p.to_string_lossy().to_string());
-
-        let process_run_time = (Utc::now() - process_start_time).num_milliseconds().max(0) as u64;
-
-        ProcessProperties::Full(Box::new(FullProcessProperties {
-            tool_name: display_name,
-            tool_pid: proc.pid().as_u32().to_string(),
-            tool_parent_pid: proc.parent().unwrap_or(0.into()).to_string(),
-            tool_binary_path: proc
-                .exe()
-                .unwrap_or_else(|| Path::new(""))
-                .as_os_str()
-                .to_str()
-                .unwrap_or("")
-                .to_string(),
-            tool_cmd: proc.cmd().join(" "),
-            start_timestamp: process_start_time.to_rfc3339(),
-            process_cpu_utilization: proc.cpu_usage(),
-            process_run_time, // time in milliseconds
-            process_disk_usage_read_total: proc.disk_usage().total_read_bytes,
-            process_disk_usage_write_total: proc.disk_usage().total_written_bytes,
-            process_disk_usage_read_last_interval: proc.disk_usage().read_bytes,
-            process_disk_usage_write_last_interval: proc.disk_usage().written_bytes,
-            process_memory_usage: proc.memory(),
-            process_memory_virtual: proc.virtual_memory(),
-            process_status: process_status_to_string(&proc.status()),
-            container_id,
-            job_id,
-            working_directory,
-            trace_id,
-        }))
-    }
-
-    /// Extracts environment variables related to containerization, jobs, and tracing
-    fn extract_process_environment_variables(
-        proc: &Process,
-    ) -> (Option<String>, Option<String>, Option<String>) {
-        let mut container_id = None;
-        let mut job_id = None;
-        let mut trace_id = None;
-
-        // Try to read environment variables
-        for env_var in proc.environ() {
-            if let Some((key, value)) = env_var.split_once('=') {
-                match key {
-                    "AWS_BATCH_JOB_ID" => job_id = Some(value.to_string()),
-                    "HOSTNAME" => container_id = Some(value.to_string()),
-                    "TRACER_TRACE_ID" => trace_id = Some(value.to_string()),
-                    _ => continue,
-                }
-            }
-        }
-
-        (container_id, job_id, trace_id)
     }
 
     /// Polls and updates metrics for all monitored processes
@@ -738,5 +657,4 @@ impl ProcessManager {
 
         Ok(())
     }
-
 }
