--- src/extracts/src/process_watcher/utils.rs
@@ -1,6 +1,0 @@
-use super::ProcessState;
-use std::collections::HashMap;
-use std::sync::Arc;
-use tokio::sync::RwLock;
-use tracer_common::types::ebpf_trigger::{ExitReason, OutOfMemoryTrigger, ProcessEndTrigger};
-use tracing::debug;
